+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bugfixes
 
 - Fix forcing of HTTPS via user configuration
+- Fix inability to save Gauge Refresh Period option ([#376](https://github.com/kizniche/mycodo/issues/376))
 
 
 ## 5.5.6 (2018-01-05)